requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1
django-heroku==0.3.1
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.2.0
